9 PM Update

+++ src/edu/grinnell/csc207/nikakath/hw2/TwoB.java
@@ -3,9 +3,13 @@ package edu.grinnell.csc207.nikakath.hw2;
 public class TwoB {
     /**
      * Compute the average of two integers.  Rounds toward zero if the
-     * average is not a whole number.
+     * average is not a whole number.  The original version of this method
+     * did not accurately calculate the averages of very large values
+     * due to dividing integer type values.
      */
     public static int average(int left, int right) {
-        return (left + right) / 2;
+        return (int) (((double)left + (double)right) / 2);
     } // average(int,int)
 } // class TwoB
+
+
